Move types to file
