+++ vercel/api/item.ts
@@ -1,3 +1,5 @@
+import * as i from './_types';
+
 export const config = {
   runtime: 'experimental-edge',
 };
@@ -10,7 +12,7 @@ export default async function handler(req: Request, res: Response) {
   console.info(`Fetching item '${query.get('id')}' for '${serverSlug}' (${factionSlug})`);
 
   const url = `https://api.nexushub.co/wow-classic/v1/items/${serverSlug}-${factionSlug}/${query.get('id')}`;
-  const result = await (await fetch(url)).json() as NexusHub.ItemsResponse | NexusHub.ErrorResponse;
+  const result = await (await fetch(url)).json() as i.NexusHub.ItemsResponse | i.NexusHub.ErrorResponse;
 
   if ('error' in result) {
     const code = result.error ? 500 : 404;
@@ -46,7 +48,7 @@ function getFactionSlug(faction = '') {
   return faction.toLowerCase();
 }
 
-function convertToCoins(rawPrice: number = 0, amount = 1): PriceObject {
+function convertToCoins(rawPrice: number = 0, amount = 1): i.PriceObject {
   const multiPrice = rawPrice * amount;
   const gold = Math.floor(multiPrice / 10000) || 0;
   const silver = Math.floor(multiPrice % 10000 / 100) || 0;
@@ -60,8 +62,8 @@ function convertToCoins(rawPrice: number = 0, amount = 1): PriceObject {
   };
 }
 
-function nexushubToItemResponse(data: NexusHub.ItemsResponse, amount = 1): ItemResponse {
-  const transformedData: ItemResponse = {
+function nexushubToItemResponse(data: i.NexusHub.ItemsResponse, amount = 1): i.ItemResponse {
+  const transformedData: i.ItemResponse = {
     ...data,
     tooltip: undefined,
     tags: data.tags.join(','),
@@ -88,92 +90,3 @@ function nexushubToItemResponse(data: NexusHub.ItemsResponse, amount = 1): ItemR
 
   return transformedData;
 }
-
-
-/** TYPES */
-type Query = {
-  id: string;
-  server_name: string;
-  faction: string;
-  amount?: string;
-};
-
-namespace NexusHub {
-  interface Tooltip {
-    label: string;
-    format: string;
-  }
-
-  interface Current {
-    historicalValue: number;
-    marketValue: number;
-    minBuyout: number;
-    numAuctions: number;
-    quantity: number;
-  }
-
-  interface Previous {
-    marketValue: number;
-    minBuyout: number;
-    quantity: number;
-    historicalValue: number;
-    numAuctions: number;
-  }
-
-  interface Stats {
-    lastUpdated: Date_ISO_8601;
-    current: Current | null;
-    previous: Previous | null;
-  }
-
-  export interface ItemsResponse {
-    server: string;
-    itemId: NumberString;
-    name: string;
-    uniqueName: string;
-    icon: string;
-    tags: string[];
-    requiredLevel: NumberString;
-    itemLevel: NumberString;
-    sellPrice: NumberString;
-    vendorPrice?: NumberString;
-    tooltip: Tooltip[];
-    itemLink: string;
-    stats: Stats;
-  }
-
-  export interface ErrorResponse {
-    error: string;
-    reason: string;
-  }
-}
-
-type Date_ISO_8601 = string;
-type NumberString = string;
-
-type PriceObject = string | {
-  gold: NumberString;
-  silver: NumberString;
-  copper: NumberString;
-  raw: NumberString;
-};
-
-type PriceSnapshot = {
-  marketValue: PriceObject;
-  historicalValue: PriceObject;
-  minimumBuyout: PriceObject;
-  numAuctions: NumberString;
-  quantity: NumberString;
-};
-
-type ItemResponse = Omit<NexusHub.ItemsResponse, 'stats' | 'tags' | 'tooltip'> & {
-  uri: string;
-  amount: NumberString;
-  stats: {
-    lastUpdated: Date_ISO_8601;
-    current: PriceSnapshot;
-    previous: PriceSnapshot;
-  };
-  tags: string;
-  tooltip: undefined;
-};
